Move DropInViewModel constructor argument to DropInViewModelFactory
COAND-679
