+++ drop-in/src/main/java/com/adyen/checkout/dropin/ui/viewmodel/DropInViewModelFactory.kt
@@ -12,12 +12,33 @@ import androidx.activity.ComponentActivity
 import androidx.lifecycle.AbstractSavedStateViewModelFactory
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
+import com.adyen.checkout.components.analytics.AnalyticsMapper
+import com.adyen.checkout.components.analytics.AnalyticsSource
+import com.adyen.checkout.components.analytics.DefaultAnalyticsRepository
+import com.adyen.checkout.components.api.AnalyticsService
+import com.adyen.checkout.components.api.OrderStatusService
+import com.adyen.checkout.components.repository.OrderStatusRepository
+import com.adyen.checkout.core.api.HttpClientFactory
+import com.adyen.checkout.dropin.DropInConfiguration
 
 internal class DropInViewModelFactory(
     activity: ComponentActivity
 ) : AbstractSavedStateViewModelFactory(activity, activity.intent.extras) {
     override fun <T : ViewModel> create(key: String, modelClass: Class<T>, handle: SavedStateHandle): T {
+        val dropInConfiguration: DropInConfiguration = requireNotNull(handle[DropInViewModel.DROP_IN_CONFIGURATION_KEY])
+        val packageName: String = requireNotNull(handle[DropInViewModel.PACKAGE_NAME_KEY])
+
+        val httpClient = HttpClientFactory.getHttpClient(dropInConfiguration.environment)
+        val orderStatusRepository = OrderStatusRepository(OrderStatusService(httpClient))
+        val analyticsRepository = DefaultAnalyticsRepository(
+            packageName = packageName,
+            locale = dropInConfiguration.shopperLocale,
+            source = AnalyticsSource.DropIn(),
+            analyticsService = AnalyticsService(httpClient),
+            analyticsMapper = AnalyticsMapper(),
+        )
+
         @Suppress("UNCHECKED_CAST")
-        return DropInViewModel(handle) as T
+        return DropInViewModel(handle, orderStatusRepository, analyticsRepository) as T
     }
 }
